fix(playerbot): Move CoordinationMetrics and fix BotLifecycleManager duplicates
- Moved CoordinationMetrics struct to IInstanceCoordination.h
- Removed duplicate BotLifecycleManager destructor and copy constructor
- All coordination metrics now defined in interface headers
