+++ src/modules/Playerbot/Core/DI/Interfaces/IInstanceCoordination.h
@@ -14,6 +14,7 @@
 #include "GameTime.h"
 #include <string>
 #include <vector>
+#include <atomic>
 
 // Forward declarations
 class Player;
@@ -22,9 +23,6 @@ class Map;
 
 namespace Playerbot
 {
-
-// Forward declarations for nested types
-struct CoordinationMetrics;
 
 // InstanceProgress definition (needs full definition for return by value)
 struct InstanceProgress
@@ -46,6 +44,31 @@ struct InstanceProgress
         , mapId(mId), startTime(GameTime::GetGameTimeMS()), currentCheckpoint(0)
         , progressPercentage(0.0f), estimatedCompletionTime(2700000) // 45 minutes
         , isOnTrack(true) {}
+};
+
+// CoordinationMetrics definition (needs full definition for return by value)
+struct CoordinationMetrics
+{
+    ::std::atomic<uint32> coordinationEvents{0};
+    ::std::atomic<uint32> successfulCoordinations{0};
+    ::std::atomic<uint32> coordinationFailures{0};
+    ::std::atomic<float> averageResponseTime{2000.0f}; // 2 seconds
+    ::std::atomic<float> groupSynchronization{0.9f}; // 90% sync rate
+    ::std::atomic<float> movementEfficiency{0.85f};
+    ::std::atomic<uint32> formationBreaks{0};
+    ::std::atomic<uint32> communicationEvents{0};
+
+    void Reset() {
+        coordinationEvents = 0; successfulCoordinations = 0; coordinationFailures = 0;
+        averageResponseTime = 2000.0f; groupSynchronization = 0.9f;
+        movementEfficiency = 0.85f; formationBreaks = 0; communicationEvents = 0;
+    }
+
+    float GetCoordinationSuccessRate() const {
+        uint32 total = coordinationEvents.load();
+        uint32 successful = successfulCoordinations.load();
+        return total > 0 ? (float)successful / total : 0.0f;
+    }
 };
 
 /**
