+++ src/modules/Playerbot/Dungeon/InstanceCoordination.h
@@ -92,31 +92,7 @@ public:
     void CoordinateEmergencyEvacuation(Group* group) override;
     void HandlePlayerIncapacitation(Group* group, Player* incapacitatedPlayer) override;
 
-    // Performance optimization
-    struct CoordinationMetrics
-    {
-        ::std::atomic<uint32> coordinationEvents{0};
-        ::std::atomic<uint32> successfulCoordinations{0};
-        ::std::atomic<uint32> coordinationFailures{0};
-        ::std::atomic<float> averageResponseTime{2000.0f}; // 2 seconds
-        ::std::atomic<float> groupSynchronization{0.9f}; // 90% sync rate
-        ::std::atomic<float> movementEfficiency{0.85f};
-        ::std::atomic<uint32> formationBreaks{0};
-        ::std::atomic<uint32> communicationEvents{0};
-
-        void Reset() {
-            coordinationEvents = 0; successfulCoordinations = 0; coordinationFailures = 0;
-            averageResponseTime = 2000.0f; groupSynchronization = 0.9f;
-            movementEfficiency = 0.85f; formationBreaks = 0; communicationEvents = 0;
-        }
-
-        float GetCoordinationSuccessRate() const {
-            uint32 total = coordinationEvents.load();
-            uint32 successful = successfulCoordinations.load();
-            return total > 0 ? (float)successful / total : 0.0f;
-        }
-    };
-
+    // Performance optimization (CoordinationMetrics defined in IInstanceCoordination.h interface)
     CoordinationMetrics GetGroupCoordinationMetrics(uint32 groupId) override;
     CoordinationMetrics GetGlobalCoordinationMetrics() override;
